Promoter page and liquid can be double

+++ promo-backend/src/main/java/hu/jager/promo_backend/dto/AddStockRequest.java
@@ -4,5 +4,5 @@ import lombok.Data;
 
 @Data
 public class AddStockRequest {
-    private int addedQuantity;
+    private double addedQuantity;
 }
